More infer clean-ups

+++ src/test/test_scrappie_convolution.c
@@ -375,7 +375,7 @@ void compare_convolution_to_baseline(const_scrappie_matrix input,
     CU_ASSERT_PTR_NOT_NULL_FATAL(res_odd);
     CU_ASSERT_PTR_NOT_NULL_FATAL(res_odd_base.elt);
 
-    for (int i = 0; i < input_base.len; ++i) {
+    for (size_t i = 0; i < input_base.len; ++i) {
         CU_ASSERT_DOUBLE_EQUAL(res_odd_base.elt[i], res_odd->data.f[i * 4],
                                test_conv_tol);
     }
